% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/add.csc.R
\name{add.csc.metaMDS}
\alias{add.csc.metaMDS}
\title{Adds circle of significant contribution onto ordination diagram}
\usage{
add.csc.metaMDS(ord, x, arrow.mul, at = c(0, 0), add = T, choices = c(1,
  2), ci = 0.95, ...)
}
\description{
Adds circle of significant contribution onto ordination diagram
}
\examples{
library (weimea)
library (vegan)
data (vltava)
env <- vltava$env$pH
#add randomly generated variables
 set.seed (1234)
 nmds <- metaMDS (vltava$spe)
env <- matrix (rnorm (8*97), ncol = 8, dimnames = list (NULL, names (vltava$env)[c(1:6, 11:12)]))
ordiplot (nmds, display = 'si')
for (gr in 1:4) ordiellipse (nmds, groups = vltava$group$GROUP, show.groups = gr, col = gr)
ef <- envfit (nmds, env)
add.csc.metaMDS (ord = nmds, x = ef, fg = 'grey', lwd = 2, ci = 0.95)
plot (ef)
}

